requirements.txt
# 历史文本项目微服务基础架构依赖

# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
asyncpg==0.29.0
psycopg2-binary==2.9.7
databases[postgresql]==0.8.0

# Redis
aioredis==2.0.1

# 服务发现
python-consul2==0.1.5

# HTTP客户端
aiohttp==3.9.1
httpx==0.25.2

# 数据模型
pydantic==2.5.0
pydantic-settings==2.1.0

# 异步支持
asyncio-mqtt==0.16.1

# 日志
loguru==0.7.2

# 监控
psutil==5.9.6

# 测试相关
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# 开发工具
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Docker相关（用于集成测试）
docker==6.1.3

# 其他工具
python-dotenv==1.0.0
typer==0.9.0